prenotazione

+++ Esercitazione4/src/main/java/com/epicode/spring/repository/UtenteDAO.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.epicode.spring.model.Utente;
 
-public interface UtenteDAO extends CrudRepository<Utente, Integer> {
+public interface UtenteDAO extends CrudRepository<Utente, Long> {
 
 }
